+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v2/validator/PostAggregateAndProofsV2Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright Consensys Software Inc., 2022
+ * Copyright Consensys Software Inc., 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -11,38 +11,58 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.beaconrestapi.handlers.v1.validator;
+package tech.pegasys.teku.beaconrestapi.handlers.v2.validator;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.when;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_BAD_REQUEST;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
+import static tech.pegasys.teku.infrastructure.http.RestApiConstants.HEADER_CONSENSUS_VERSION;
+import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.getRequestBodyFromMetadata;
 import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.verifyMetadataEmptyResponse;
 import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.verifyMetadataErrorResponse;
+import static tech.pegasys.teku.spec.SpecMilestone.ELECTRA;
+import static tech.pegasys.teku.spec.SpecMilestone.PHASE0;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.google.common.io.Resources;
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestTemplate;
 import tech.pegasys.teku.beaconrestapi.AbstractMigratedBeaconHandlerTest;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.http.HttpErrorResponse;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.SpecMilestone;
+import tech.pegasys.teku.spec.TestSpecContext;
+import tech.pegasys.teku.spec.TestSpecInvocationContextProvider;
 import tech.pegasys.teku.spec.datastructures.operations.SignedAggregateAndProof;
+import tech.pegasys.teku.spec.datastructures.operations.versions.electra.AttestationElectra;
+import tech.pegasys.teku.spec.datastructures.operations.versions.phase0.AttestationPhase0;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionCache;
+import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.teku.validator.api.SubmitDataError;
 
-public class PostAggregateAndProofsTest extends AbstractMigratedBeaconHandlerTest {
+@TestSpecContext(milestone = {PHASE0, ELECTRA})
+public class PostAggregateAndProofsV2Test extends AbstractMigratedBeaconHandlerTest {
+
+  private SpecMilestone specMilestone;
 
   @BeforeEach
-  public void beforeEach() {
-    setHandler(new PostAggregateAndProofs(validatorDataProvider, new SchemaDefinitionCache(spec)));
+  public void beforeEach(final TestSpecInvocationContextProvider.SpecContext specContext) {
+    spec = specContext.getSpec();
+    dataStructureUtil = new DataStructureUtil(spec);
+    specMilestone = specContext.getSpecMilestone();
+    setHandler(
+        new PostAggregateAndProofsV2(validatorDataProvider, new SchemaDefinitionCache(spec)));
   }
 
-  @Test
+  @TestTemplate
   public void shouldReturnSuccessWhenSendAggregateAndProofSucceeds() throws Exception {
     final SignedAggregateAndProof signedAggregateAndProof =
         dataStructureUtil.randomSignedAggregateAndProof();
@@ -57,7 +77,7 @@ public class PostAggregateAndProofsTest extends AbstractMigratedBeaconHandlerTes
     assertThat(request.getResponseBody()).isNull();
   }
 
-  @Test
+  @TestTemplate
   public void shouldReturnBadRequestWhenErrorsReturned() throws Exception {
     final SignedAggregateAndProof signedAggregateAndProof =
         dataStructureUtil.randomSignedAggregateAndProof();
@@ -76,18 +96,48 @@ public class PostAggregateAndProofsTest extends AbstractMigratedBeaconHandlerTes
     assertThat(request.getResponseBody()).isEqualTo(expected);
   }
 
-  @Test
+  @TestTemplate
+  void shouldReadRequestBody() throws IOException {
+    final String data = getExpectedResponseAsJson(specMilestone);
+    final Object requestBody =
+        getRequestBodyFromMetadata(
+            handler, Map.of(HEADER_CONSENSUS_VERSION, specMilestone.name()), data);
+    assertThat(requestBody).isInstanceOf(List.class);
+    assertThat(((List<?>) requestBody).get(0)).isInstanceOf(SignedAggregateAndProof.class);
+    if (specMilestone.isGreaterThanOrEqualTo(ELECTRA)) {
+      assertThat(
+              ((SignedAggregateAndProof) ((List<?>) requestBody).get(0))
+                  .getMessage()
+                  .getAggregate())
+          .isInstanceOf(AttestationElectra.class);
+    } else {
+      assertThat(
+              ((SignedAggregateAndProof) ((List<?>) requestBody).get(0))
+                  .getMessage()
+                  .getAggregate())
+          .isInstanceOf(AttestationPhase0.class);
+    }
+  }
+
+  @TestTemplate
   void metadata_shouldHandle400() throws IOException {
     verifyMetadataErrorResponse(handler, SC_BAD_REQUEST);
   }
 
-  @Test
+  @TestTemplate
   void metadata_shouldHandle500() throws JsonProcessingException {
     verifyMetadataErrorResponse(handler, SC_INTERNAL_SERVER_ERROR);
   }
 
-  @Test
+  @TestTemplate
   void metadata_shouldHandle200() {
     verifyMetadataEmptyResponse(handler, SC_OK);
+  }
+
+  private String getExpectedResponseAsJson(final SpecMilestone specMilestone) throws IOException {
+    final String fileName =
+        String.format("postAggregateAndProofsRequestBody%s.json", specMilestone.name());
+    return Resources.toString(
+        Resources.getResource(PostAggregateAndProofsV2Test.class, fileName), UTF_8);
   }
 }